+++ sdk/core/azure-core/inc/azure/core/context.hpp
@@ -40,7 +40,44 @@ namespace Azure { namespace Core {
   };
 
   /**
-   * @brief A context is a node within a tree that represents deadlines and key/value pairs.
+   * @brief A context is a node within a unidirectional tree that represents deadlines and key/value
+   * pairs.
+   *
+   * All Azure Service Client operations which can take a long time to complete accept a Context
+   * object. The Context object allows the caller to cancel the operation if it is taking too long,
+   * or to ensure that the operation completes before a specific deadline. This allows an
+   * application to apply timeouts to operations, or to cancel operations that are no longer needed.
+   *
+   * By default, each newly created Context object is a child of the singleton ApplicationContext
+   * object.
+   * 
+   * After cancelling a Context, all operations which have the cancelled context as a parent context will be cancelled.
+   * 
+   * Cancellation is indicated by throwing an Azure::Core::OperationCancelledException from the
+   * operation.
+   *
+   * Context objects support three operations to create new contexts:
+   * - WithDeadline(DateTime): creates a new child context with a deadline.
+   * - WithCancel(): creates a new child context that can be independently cancelled.
+   * - WithValue(Key, T): creates a new child context with a key/value pair.
+   *
+   * Context objects support two operations to retrieve data:
+   * - GetDeadline(): gets the deadline for the context.
+   * - TryGetValue(Key, T): gets the value associated with a key.
+   *
+   * Context objects support two operations to manage the context:
+   * - Cancel(): cancels the context.
+   * - IsCancelled(): checks if the context is cancelled.
+   *
+   * Context objects support one operation to throw if the context is cancelled:
+   * - ThrowIfCancelled(): throws an OperationCancelledException if the context is cancelled.
+   *
+   * Context objects support one operation to create a new root context:
+   * - CreateNewRoot(): creates a new root context.
+   *
+   * Root Context objects support one operation to reset the context:
+   * - Reset(): resets the context to its original state.
+   *
    */
   class Context final {
   public:
@@ -96,9 +133,8 @@ namespace Azure { namespace Core {
         return DateTime(DateTime::time_point(DateTime::duration(dtRepresentation)));
       }
 
-      /** Creates a new ContextSharedState object with no deadline and no value.
-       *
-       *
+      /**
+       * @brief Creates a new ContextSharedState object with no deadline and no value.
        */
       explicit ContextSharedState()
           : Deadline(ToDateTimeRepresentation((DateTime::max)())), Value(nullptr)
@@ -109,8 +145,9 @@ namespace Azure { namespace Core {
       {
       }
 
-      /** Create a new ContextSharedState from another ContextSharedState with no deadline and no
-       * value
+      /**
+       * @brief Create a new ContextSharedState from another ContextSharedState with no deadline and
+       * no value
        */
       explicit ContextSharedState(const std::shared_ptr<ContextSharedState>& parent)
           : Parent(parent), Deadline(ToDateTimeRepresentation((DateTime::max)())), Value(nullptr)
@@ -121,7 +158,8 @@ namespace Azure { namespace Core {
       {
       }
 
-      /** Create a new ContextSharedState from another ContextSharedState with a deadline.
+      /**
+       * @brief Create a new ContextSharedState from another ContextSharedState with a deadline.
        *
        * @param parent The parent context to create a child context from.
        * @param deadline The deadline for the new context.
@@ -138,8 +176,9 @@ namespace Azure { namespace Core {
       {
       }
 
-      /** Create a new ContextSharedState from another ContextSharedState with a deadline and a key
-       * value pair.
+      /**
+       * @brief Create a new ContextSharedState from another ContextSharedState with a deadline and
+       * a key value pair.
        *
        * @tparam T The type of the value to be stored with the key.
        * @param parent The parent context to create a child context from.
@@ -182,7 +221,8 @@ namespace Azure { namespace Core {
     {
     }
 
-    /** Copies a context.
+    /**
+     * @brief Copies a context.
      *
      * This operation copies one context to another. Context objects are copied by reference, so the
      * new context will share the same state as the original context. This also means that
@@ -193,7 +233,8 @@ namespace Azure { namespace Core {
      */
     Context(Context const&) = default;
 
-    /** Assigns a context.
+    /**
+     * @brief Assigns a context.
      *
      * This operation assigns one context to another. Context objects are copied by reference, so
      * the new context will share the same state as the original context. This also means that
@@ -205,7 +246,8 @@ namespace Azure { namespace Core {
      */
     Context& operator=(Context const& other) = default;
 
-    /** Moves a context.
+    /**
+     * @brief Moves a context.
      *
      * This operation moves one context to another.
      *
@@ -214,7 +256,8 @@ namespace Azure { namespace Core {
      */
     Context(Context&& other) = default;
 
-    /** Moves a context.
+    /**
+     * @brief Moves a context.
      *
      * This operation moves one context to another.
      *
@@ -247,7 +290,7 @@ namespace Azure { namespace Core {
      *
      * @return A child of the current context which can be independently cancelled.
      */
-    Context WithCancellation() const
+    Context WithCancel() const
     {
       return Context{std::make_shared<ContextSharedState>(m_contextSharedState)};
     }
@@ -294,7 +337,7 @@ namespace Azure { namespace Core {
      */
     template <class T> bool TryGetValue(Key const& key, T& outputValue) const
     {
-      for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
+      for (std::shared_ptr<ContextSharedState> ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
       {
         if (ptr->Key == key)
         {
@@ -310,7 +353,15 @@ namespace Azure { namespace Core {
       return false;
     }
 
-    /** @brief Cancels the context. All operations which share this Context will be cancelled.
+    /**
+     * @brief Cancels the context. All operations which share this Context will be cancelled.
+     *
+     * @note Cancellation of a Context is a best-faith effort. Because of the synchronous nature of
+     * Azure C++ SDK APIs, it is possible that the operation will not be cancelled immediately. Each
+     * operation explicitly checks the context's state to determine if it has been cancelled,
+     * those checks may not happen immediately, or at all.
+     *
+     * @note Cancellation of a Context must be explicitly undone by
      *
      */
     void Cancel()
@@ -319,7 +370,34 @@ namespace Azure { namespace Core {
           = ContextSharedState::ToDateTimeRepresentation((DateTime::min)());
     }
 
-    /** @brief Checks if the context is cancelled.
+    /**
+     * @brief Resets the context to its original state.
+     *
+     * @note This method is only available for root Context objects.
+     *
+     * @note This method will abandon the current shared state and reset it to a new root.
+     *
+     * @note This method can only be called on a cancelled context.
+     *
+     */
+    void Reset()
+    {
+      if (m_contextSharedState->Parent)
+      {
+        throw std::runtime_error("Reset is only available for root Context objects.");
+      }
+
+      if (!IsCancelled())
+      {
+        throw std::runtime_error("Cannot reset a non-cancelled context.");
+      }
+
+      // Abandon the current shared state and reset it to the root.
+      m_contextSharedState = std::make_shared<ContextSharedState>();
+    }
+
+    /**
+     * @brief Checks if the context is cancelled.
      * @return `true` if this context is cancelled; otherwise, `false`.
      */
     bool IsCancelled() const { return GetDeadline() < std::chrono::system_clock::now(); }
@@ -349,9 +427,9 @@ namespace Azure { namespace Core {
      * @note Cancelling the `ApplicationContext` will cancel ALL operations currently active in the
      * process, unless they were created with Azure::Core::Context::CreateNewRoot().
      *
-     * @note If the `ApplicationContext` is cancelled, all subsequent operations will be cancelled.
-     * To reset the `ApplicationContext` to a non-cancelled state, use
-     * Azure::Core::Context::CreateNewRoot().
+     * @note If the `ApplicationContext` is cancelled, all current and future operations will be
+     * cancelled. To reset the `ApplicationContext` to a non-cancelled state, call
+     * Azure::Core::Context::ApplicationContext.Reset().
      *
      */
     static AZ_CORE_DLLEXPORT Context ApplicationContext;