+++ sdk/core/azure-core/CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### Breaking Changes
 
-- All newly created `Azure::Core::Context` derive from a shared `Azure::Core::Context::ApplicationContext` object. 
-  - This change is not expected to impact most users, but it may affect users which cancelled a copy of the Azure::Core::Context::ApplicationContext object.
+- All newly created `Azure::Core::Context` derive from a shared `Azure::Core::Context::ApplicationContext` object. Before this change, each `Azure::Core::Context` object created a new root `Context`.
+  - This change is not expected to impact most users, but it may affect users which had previously cancelled a copy of the Azure::Core::Context::ApplicationContext object.
 
 ### Bugs Fixed
 
@@ -18,7 +18,9 @@
 - [[#5622]](https://github.com/Azure/azure-sdk-for-cpp/pull/5622) Documentation fix for building the SDK with specific OpenSSL version.  (A community contribution, courtesy of _[ByteYue](https://github.com/ByteYue)_)
 - By default, Azure::Core::Context has been updated to be a child of `Azure::Core::Context::ApplicationContext`. In addition, the following new functions were added:
   - `Azure::Core::Context::CreateNewRoot()` - this creates a new root context. This is useful when you want to create a new context that will not be affected by the cancellation of `Azure::Core::Context::ApplicationContext`
-  - `Azure::Core::Context::WithCancellation()` - this creates a new independently cancellable `Context` as a child of an existing Context.
+    - It is also useful to restore the previous behavior of each `Azure::Core::Context` being created as an independant root.
+  - `Azure::Core::Context::WithCancel()` - this creates a new independently cancellable `Context` as a child of an existing Context.
+  - `Azure::Core::Context::Reset()` - This resets a Root context to its initial state. This is useful when you want to reset the cancellation state of `Azure::Core::Context::ApplicationContext`.
 - Made `ApplicationContext` mutable - this undoes an unintended breaking change introduced in version 1.8.1.
 
 ### Acknowledgments